Create MovingSquaresWithKeyAndMouse.html
Using WebGL and GLSL, this program shows a number of moving red squares that bounce off the edges of the canvas. The canvas fills the entire content area of the web browser. The squares also respond to the mouse: If you left-click or left-click-and-drag on the canvas, all of the square will head towards the position of the mouse. If you shift-left-click, the data for the points is reinitialized, so they start back at the center. You can pause and restart the animation by hitting the space bar. The squares are in fact all part of one WebGL primitive of type gl.POINTS. Each square corresponds to one of the vertices of the primitive. It is possible to draw the squares in different colors, to draw disks instead of squares, and so on, along with lines connecting the points.